[frontend] create reusable component for findings list
