+++ openbas-front/src/admin/components/atomic_testings/atomic_testing/AtomicTestingFindings.tsx
@@ -1,47 +1,23 @@
-import { List, ListItem, ListItemIcon, ListItemText } from '@mui/material';
-import { type CSSProperties, type FunctionComponent, useContext, useState } from 'react';
+import { type FunctionComponent, useContext, useState } from 'react';
 import { useParams, useSearchParams } from 'react-router';
-import { makeStyles } from 'tss-react/mui';
 
 import { searchFindingsForInjects } from '../../../../actions/findings/finding-actions';
 import { initSorting } from '../../../../components/common/queryable/Page';
 import PaginationComponentV2 from '../../../../components/common/queryable/pagination/PaginationComponentV2';
 import { buildSearchPagination } from '../../../../components/common/queryable/QueryableUtils';
-import SortHeadersComponentV2 from '../../../../components/common/queryable/sort/SortHeadersComponentV2';
-import useBodyItemsStyles from '../../../../components/common/queryable/style/style';
 import { useQueryableWithLocalStorage } from '../../../../components/common/queryable/useQueryableWithLocalStorage';
-import FindingIcon from '../../../../components/FindingIcon';
-import ItemTags from '../../../../components/ItemTags';
-import ItemTargets from '../../../../components/ItemTargets';
 import {
   type Finding,
   type InjectResultOverviewOutput, type TargetSimple,
 } from '../../../../utils/api-types';
+import FindingList from '../../findings/FindingList';
 import {
   InjectResultOverviewOutputContext,
   type InjectResultOverviewOutputContextType,
 } from '../InjectResultOverviewOutputContext';
 
-const useStyles = makeStyles()(() => ({
-  itemHead: { textTransform: 'uppercase' },
-  item: { height: 50 },
-}));
-
-const inlineStyles: Record<string, CSSProperties> = {
-  finding_type: { width: '20%' },
-  finding_field: { width: '20%' },
-  finding_value: {
-    width: '30%',
-    cursor: 'default',
-  },
-  finding_assets: { width: '10%' },
-  finding_tags: { width: '20%' },
-};
-
 const AtomicTestingFindings: FunctionComponent = () => {
-  const { classes } = useStyles();
   const { injectId } = useParams() as { injectId: InjectResultOverviewOutput['inject_id'] };
-  const bodyItemsStyles = useBodyItemsStyles();
 
   const availableFilterNames = ['finding_type'];
   const { injectResultOverviewOutput } = useContext<InjectResultOverviewOutputContextType>(InjectResultOverviewOutputContext);
@@ -64,40 +40,6 @@ const AtomicTestingFindings: FunctionComponent = () => {
     textSearch: search,
   }));
 
-  // Headers
-  const headers = [
-    {
-      field: 'finding_type',
-      label: 'Type',
-      isSortable: true,
-      value: (finding: Finding) => finding.finding_type,
-    },
-    {
-      field: 'finding_field',
-      label: 'Key',
-      isSortable: true,
-      value: (finding: Finding) => finding.finding_field,
-    },
-    {
-      field: 'finding_value',
-      label: 'Value',
-      isSortable: true,
-      value: (finding: Finding) => finding.finding_value,
-    },
-    {
-      field: 'finding_assets',
-      label: 'Asset',
-      isSortable: false,
-      value: (finding: Finding) => <ItemTargets targets={(finding.finding_assets || []).map(assetId => assetsMap.get(assetId)) as TargetSimple[]} />,
-    },
-    {
-      field: 'finding_tags',
-      label: 'Tags',
-      isSortable: false,
-      value: (finding: Finding) => <ItemTags variant="list" tags={finding.finding_tags} />,
-    },
-  ];
-
   return (
     <>
       <PaginationComponentV2
@@ -108,52 +50,7 @@ const AtomicTestingFindings: FunctionComponent = () => {
         availableFilterNames={availableFilterNames}
         queryableHelpers={queryableHelpers}
       />
-      <List>
-        <ListItem
-          classes={{ root: classes.itemHead }}
-          style={{ padding: 0 }}
-        >
-          <ListItemIcon />
-          <ListItemText
-            primary={(
-              <SortHeadersComponentV2
-                headers={headers}
-                inlineStylesHeaders={inlineStyles}
-                sortHelpers={queryableHelpers.sortHelpers}
-              />
-            )}
-          />
-        </ListItem>
-        {findings.map((finding: Finding) => (
-          <ListItem
-            key={finding.finding_id}
-            classes={{ root: classes.item }}
-            divider={true}
-            disablePadding={true}
-          >
-            <ListItemIcon>
-              <FindingIcon findingType={finding.finding_type} tooltip={true} />
-            </ListItemIcon>
-            <ListItemText
-              primary={(
-                <div style={bodyItemsStyles.bodyItems}>
-                  {headers.map(header => (
-                    <div
-                      key={header.field}
-                      style={{
-                        ...bodyItemsStyles.bodyItem,
-                        ...inlineStyles[header.field],
-                      }}
-                    >
-                      {header.value(finding)}
-                    </div>
-                  ))}
-                </div>
-              )}
-            />
-          </ListItem>
-        ))}
-      </List>
+      <FindingList findings={findings} sortHelpers={queryableHelpers.sortHelpers} assetsMap={assetsMap} />
     </>
   );
 };
